kahesh estekak & scorebord

+++ src/rasma.h
@@ -8,5 +8,5 @@
 #include "structha.h"
 #include <SDL2_gfxPrimitives.h>
 void rasmtir(double snake_radiuse,tanks tank[],SDL_Renderer *renderer,int step,Map map);
-jatank(int adad ,tanks tank[],int tool,int rasm);
+jatank(int adad ,tanks tank[], double tool, double rasm);
 #endif //PROJECT_RASMA_H
